refactor: make SelectMenuPaginator specific to the bot
