--- src/util/if.ts
@@ -1,1 +1,0 @@
-export type If<Condition extends boolean, Then, Else> = Condition extends true ? Then : Else;
